[~] refactoring code
