+++ tvApp/src/main/java/band/effective/office/tv/repository/supernova/Talent.kt
@@ -2,7 +2,7 @@ package band.effective.office.tv.repository.supernova
 
 import band.effective.office.tv.domain.model.DomainModel
 import band.effective.office.tv.domain.model.notion.EmployeeInfoEntity
-import band.effective.office.tv.screen.supernova.model.SupernovaUserUi
+import band.effective.office.tv.screen.ratings.supernova.model.SupernovaUserUi
 
 data class Talent(
     val id: String,
