+++ tvApp/src/main/java/band/effective/office/tv/screen/ratings/RatingScreen.kt
@@ -1,6 +1,5 @@
-package band.effective.office.tv.screen.supernova
+package band.effective.office.tv.screen.ratings
 
-import band.effective.office.tv.screen.supernova.model.SupernovaUserUi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -10,19 +9,20 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import band.effective.office.tv.R
-import band.effective.office.tv.screen.components.TitleRating
-import band.effective.office.tv.screen.components.TopRating
-import band.effective.office.tv.ui.theme.EffectiveColor
 
 @Composable
-fun SupernovaScreen(
-    supernovaUsers: List<SupernovaUserUi>
+fun <T> RatingScreen(
+    users: List<T>,
+    backgroundColor: Color,
+    titlePath: Int,
+    logoPath: Int,
+    ratingTop: @Composable (List<T>) -> Unit
 ) {
     Box(
         modifier = Modifier
-            .background(EffectiveColor.supernova)
+            .background(backgroundColor)
             .fillMaxSize()
     ) {
         Column(
@@ -31,16 +31,11 @@ fun SupernovaScreen(
                 .padding(start = 60.dp, end = 120.dp)
         ) {
             TitleRating(
-                imagePath = R.drawable.supernova,
-                stringPath = R.string.supernova_title
+                imagePath = logoPath,
+                stringPath = titlePath
             )
             Spacer(modifier = Modifier.height(30.dp))
-            TopRating(users = supernovaUsers) { modifier, supernovaUi, _ ->
-                SupernovaItem(
-                    modifier = modifier,
-                    user = supernovaUi
-                )
-            }
+            ratingTop(users)
         }
     }
 }
